requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
google-generativeai==0.3.2
Flask-Session==0.5.0
cryptography==41.0.4
requests==2.31.0
werkzeug==2.3.7
python-dotenv==1.0.0
